kack rechen fehler gefixt

+++ default/body.py
@@ -7,7 +7,7 @@ Created on Sat Nov  4 20:20:31 2017
 """
 import numpy as np
 class Body(object):
-    def __init__(self, mass, radius, scale, maxrad, initdir=-1, initpos=np.array((0,0,0))):
+    def __init__(self, mass, radius, scale, maxrad, initdir=-1, initpos=np.array((0,0,0), dtype=np.float64)):
         self.mass = mass*1000
         self.radius = radius
         self.speed = 0
